/Order/Type作成、order.html修正

+++ src/main/java/com/iwamih31/KAKEIBO/Type.java
@@ -32,6 +32,6 @@ public class Type {
 
   // 並び順　0は非表示
   @Column(name = "rank", nullable = false)
-  private Integer rank;
+  private double rank;
 
 }
